requirements.txt
streamlit
PyGithub
openai